Changed lists into pluck, added Admin LTE box, icons, confirm on delete

+++ portal/resources/views/groups/index.blade.php
@@ -4,45 +4,47 @@
 @endsection
 
 @section('content')
-	<div class="row">
-	    <div class="col-lg-12 margin-tb">
-	        <div class="pull-right">
-	        	@permission('group-create')
-	            <a class="btn btn-success" href="{{ route('groups.create') }}"> Create new group</a>
+
+			
+	@component('components/box')
+		@slot('title')
+			Group overview
+		@endslot
+
+		@slot('action')
+			@permission('group-create')
+	            <a class="btn btn-success btn-xs" href="{{ route('groups.create') }}"><i class="fa fa-plus"></i> Add new group</a>
 	            @endpermission
-	        </div>
-	    </div>
-	</div>
-	@if ($message = Session::get('success'))
-		<div class="alert alert-success">
-			<p>{{ $message }}</p>
-		</div>
-	@endif
-	<table class="table table-bordered">
-		<tr>
-			<th>No</th>
-			<th>Name</th>
-			<th>Type</th>
-			<th width="280px">Action</th>
-		</tr>
-	@foreach ($groups as $key => $group)
-	<tr>
-		<td>{{ ++$i }}</td>
-		<td>{{ $group->name }}</td>
-		<td>{{ $group->type }}</td>
-		<td>
-			<a class="btn btn-info" href="{{ route('groups.show',$group->id) }}">Show</a>
-			@permission('group-edit')
-			<a class="btn btn-primary" href="{{ route('groups.edit',$group->id) }}">Edit</a>
-			@endpermission
-			@permission('group-delete')
-			{!! Form::open(['method' => 'DELETE','route' => ['groups.destroy', $group->id],'style'=>'display:inline']) !!}
-            {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-        	{!! Form::close() !!}
-        	@endpermission
-		</td>
-	</tr>
-	@endforeach
-	</table>
-	{!! $groups->render() !!}
+		@endslot
+
+		@slot('body')
+			<table class="table table-striped">
+				<tr>
+					<th>No</th>
+					<th>Name</th>
+					<th>Type</th>
+					<th width="280px">Action</th>
+				</tr>
+			@foreach ($groups as $key => $group)
+			<tr>
+				<td>{{ ++$i }}</td>
+				<td>{{ $group->name }}</td>
+				<td>{{ $group->type }}</td>
+				<td>
+					<a class="btn btn-info btn-xs" href="{{ route('groups.show',$group->id) }}" title="Show"><i class="fa fa-eye"></i></a>
+					@permission('group-edit')
+					<a class="btn btn-primary btn-xs" href="{{ route('groups.edit',$group->id) }}" title="Edit"><i class="fa fa-pencil"></i></a>
+					@endpermission
+					@permission('group-delete')
+					{!! Form::open(['method' => 'DELETE','route' => ['groups.destroy', $group->id],'style'=>'display:inline', 'onsubmit'=>'return confirm("Are you sure you want to delete group '.$group->name.'?")']) !!}
+		            {!! Form::button('<i class="fa fa-trash-o"></i>', ['type'=>'submit', 'class' => 'btn btn-danger btn-xs pull-right']) !!}
+		        	{!! Form::close() !!}
+		        	@endpermission
+				</td>
+			</tr>
+			@endforeach
+			</table>
+			{!! $groups->render() !!}
+		@endslot
+	@endcomponent
 @endsection
